added arrow renderer resolved thread deletion bug added method to enumerate shader program attributes added method to check whether shader program is enabled added support to dlip texture v coordinate when covering the screen with a texture added enum shortcuts for combinations of surface color mapping resolved filename gui bug extended picker example to 3d surfaces and multiple viewports

+++ cgv/render/textured_material.cxx
@@ -139,7 +139,7 @@ void textured_material::disable_textures(context& ctx)
 /// return pointer to ambient texture or 0 if non created
 texture* textured_material::get_texture(int ti) const
 {
-	assert(ti < textures.size());
+	assert(ti < (int)textures.size());
 	return textures[ti];
 }
